+++ docs/source/changelog.rst
@@ -5,9 +5,9 @@ Changelog
 	:local:
 	:backlinks: none
 
--------------
-Version (dev)
--------------
+--------------
+Version 2.3.17
+--------------
 * Added :py:meth:`Keyvalues.serialise() <srctools.keyvalues.Keyvalues.serialise>`, a replacement for :py:meth:`~srctools.keyvalues.Keyvalues.export`.
 * Fix `+` and `=` being parsed as part of a bare string if not the first character.
 * Fix keyvalue-type snippets causing a parse error for code coming after them.